Allow skipping of root device check on OSDs
The check to confirm the system root device has not been included in
the list of OSD storage devices does not work when using ZFS root
filesystem as this can use multiple devices in a zpool. So there
is now a variable 'osd_check_root_device' which can be set to
false to skip this check when using ZFS root disks.